+++ src/Tests/Test.Common/Models/QueueMessageA.cs
@@ -2,6 +2,11 @@ namespace Test.Common.Models
 {
     public class QueueMessageA
     {
-        
+        public string Foo { get; set; }
+
+        public QueueMessageA(string foo)
+        {
+            Foo = foo;
+        }
     }
 }
