+++ src/Tests/Test.Direct/Test.Direct.csproj
@@ -13,4 +13,12 @@
         <PackageReference Include="coverlet.collector" Version="1.2.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\..\Twino.Client.TMQ\Twino.Client.TMQ.csproj" />
+      <ProjectReference Include="..\..\Twino.MQ.Data\Twino.MQ.Data.csproj" />
+      <ProjectReference Include="..\..\Twino.MQ\Twino.MQ.csproj" />
+      <ProjectReference Include="..\..\Twino.Protocols.TMQ\Twino.Protocols.TMQ.csproj" />
+      <ProjectReference Include="..\Test.Common\Test.Common.csproj" />
+    </ItemGroup>
+
 </Project>
