--- src/Tests/Test.Persistency/UnitTest1.cs
@@ -1,13 +1,0 @@
-using System;
-using Xunit;
-
-namespace Test.Persistency
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-        }
-    }
-}
