+++ src/Tests/Test.Queues/Statuses/PullStatusTest.cs
@@ -25,24 +25,24 @@ namespace Test.Mq.Statuses
             consumer.ClientId = "consumer";
             await consumer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(consumer.IsConnected);
-            TwinoResult joined = await consumer.Channels.Join("ch-pull", true);
+            TwinoResult joined = await consumer.Channels.Join("pull-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             TmqClient producer = new TmqClient();
             await producer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(producer.IsConnected);
 
-            await producer.Queues.Push("ch-pull", MessageA.ContentType, "Hello, World!", false);
+            await producer.Queues.Push("pull-a", MessageA.ContentType, "Hello, World!", false);
             await Task.Delay(700);
 
-            Channel channel = server.Server.FindChannel("ch-pull");
+            Channel channel = server.Server.FindChannel("pull-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             Assert.NotNull(channel);
             Assert.NotNull(queue);
             Assert.Single(queue.Messages);
 
             PullRequest request = new PullRequest();
-            request.Channel = "ch-pull";
+            request.Channel = "pull-a";
             request.QueueId = MessageA.ContentType;
             request.Count = 1;
             request.ClearAfter = ClearDecision.None;
@@ -65,7 +65,7 @@ namespace Test.Mq.Statuses
             server.Initialize();
             int port = server.Start(300, 300);
 
-            Channel channel = server.Server.FindChannel("ch-pull");
+            Channel channel = server.Server.FindChannel("pull-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             Assert.NotNull(channel);
             Assert.NotNull(queue);
@@ -81,7 +81,7 @@ namespace Test.Mq.Statuses
 
             bool msgReceived = false;
             consumer.MessageReceived += (c, m) => msgReceived = true;
-            TwinoResult joined = await consumer.Channels.Join("ch-pull", true);
+            TwinoResult joined = await consumer.Channels.Join("pull-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             TmqClient producer = new TmqClient();
@@ -89,7 +89,7 @@ namespace Test.Mq.Statuses
             await producer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(producer.IsConnected);
 
-            Task<TwinoResult> taskAck = producer.Queues.Push("ch-pull", MessageA.ContentType, "Hello, World!", true);
+            Task<TwinoResult> taskAck = producer.Queues.Push("pull-a", MessageA.ContentType, "Hello, World!", true);
 
             await Task.Delay(500);
             Assert.False(taskAck.IsCompleted);
@@ -98,7 +98,7 @@ namespace Test.Mq.Statuses
 
             consumer.PullTimeout = TimeSpan.FromDays(1);
 
-            PullContainer pull = await consumer.Queues.Pull(PullRequest.Single("ch-pull", MessageA.ContentType));
+            PullContainer pull = await consumer.Queues.Pull(PullRequest.Single("pull-a", MessageA.ContentType));
             Assert.Equal(PullProcess.Completed, pull.Status);
             Assert.Equal(1, pull.ReceivedCount);
             Assert.NotEmpty(pull.ReceivedMessages);
@@ -117,19 +117,19 @@ namespace Test.Mq.Statuses
             server.Initialize();
             int port = server.Start();
 
-            var channel = server.Server.FindChannel("ch-pull");
+            var channel = server.Server.FindChannel("pull-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             queue.AddStringMessageWithId("First Message");
             queue.AddStringMessageWithId("Second Message");
 
             TmqClient client = new TmqClient();
             await client.ConnectAsync("tmq://localhost:" + port);
-            TwinoResult joined = await client.Channels.Join("ch-pull", true);
+            TwinoResult joined = await client.Channels.Join("pull-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             PullRequest request = new PullRequest
                                   {
-                                      Channel = "ch-pull",
+                                      Channel = "pull-a",
                                       QueueId = MessageA.ContentType,
                                       Count = 1,
                                       Order = !fifo.HasValue || fifo.Value ? MessageOrder.FIFO : MessageOrder.LIFO
@@ -161,19 +161,19 @@ namespace Test.Mq.Statuses
             server.Initialize();
             int port = server.Start();
 
-            var channel = server.Server.FindChannel("ch-pull");
+            var channel = server.Server.FindChannel("pull-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             for (int i = 0; i < 25; i++)
                 queue.AddStringMessageWithId("Hello, World");
 
             TmqClient client = new TmqClient();
             await client.ConnectAsync("tmq://localhost:" + port);
-            TwinoResult joined = await client.Channels.Join("ch-pull", true);
+            TwinoResult joined = await client.Channels.Join("pull-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             PullRequest request = new PullRequest
                                   {
-                                      Channel = "ch-pull",
+                                      Channel = "pull-a",
                                       QueueId = MessageA.ContentType,
                                       Count = count
                                   };
@@ -196,7 +196,7 @@ namespace Test.Mq.Statuses
             server.Initialize();
             int port = server.Start();
 
-            var channel = server.Server.FindChannel("ch-pull");
+            var channel = server.Server.FindChannel("pull-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             for (int i = 0; i < 5; i++)
             {
@@ -206,7 +206,7 @@ namespace Test.Mq.Statuses
 
             TmqClient client = new TmqClient();
             await client.ConnectAsync("tmq://localhost:" + port);
-            TwinoResult joined = await client.Channels.Join("ch-pull", true);
+            TwinoResult joined = await client.Channels.Join("pull-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             ClearDecision clearDecision = ClearDecision.None;
@@ -219,7 +219,7 @@ namespace Test.Mq.Statuses
 
             PullRequest request = new PullRequest
                                   {
-                                      Channel = "ch-pull",
+                                      Channel = "pull-a",
                                       QueueId = MessageA.ContentType,
                                       Count = count,
                                       ClearAfter = clearDecision
