+++ src/Tests/Test.Queues/Statuses/RoundRobinStatusTest.cs
@@ -39,13 +39,13 @@ namespace Test.Mq.Statuses
                 await consumer.ConnectAsync("tmq://localhost:" + port);
                 Assert.True(consumer.IsConnected);
                 consumer.MessageReceived += (c, m) => Interlocked.Increment(ref msgReceived);
-                TwinoResult joined = await consumer.Channels.Join("ch-round", true);
+                TwinoResult joined = await consumer.Channels.Join("rr-a", true);
                 Assert.Equal(TwinoResultCode.Ok, joined.Code);
             }
 
             for (int i = 0; i < 50; i++)
             {
-                await producer.Queues.Push("ch-round", MessageA.ContentType, "Hello, World!", false);
+                await producer.Queues.Push("rr-a", MessageA.ContentType, "Hello, World!", false);
                 msgSent++;
             }
 
@@ -64,10 +64,10 @@ namespace Test.Mq.Statuses
             await producer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(producer.IsConnected);
 
-            await producer.Queues.Push("ch-round", MessageA.ContentType, "Hello, World!", false);
+            await producer.Queues.Push("rr-a", MessageA.ContentType, "Hello, World!", false);
             await Task.Delay(700);
 
-            Channel channel = server.Server.FindChannel("ch-round");
+            Channel channel = server.Server.FindChannel("rr-a");
             TwinoQueue queue = channel.FindQueue(MessageA.ContentType);
             Assert.NotNull(channel);
             Assert.NotNull(queue);
@@ -79,7 +79,7 @@ namespace Test.Mq.Statuses
             await consumer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(consumer.IsConnected);
             consumer.MessageReceived += (c, m) => msgReceived = true;
-            TwinoResult joined = await consumer.Channels.Join("ch-round", true);
+            TwinoResult joined = await consumer.Channels.Join("rr-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
             await Task.Delay(800);
@@ -95,7 +95,7 @@ namespace Test.Mq.Statuses
             server.Initialize();
             int port = server.Start(300, 300);
 
-            Channel ch = server.Server.FindChannel("ch-round");
+            Channel ch = server.Server.FindChannel("rr-a");
             TwinoQueue queue = ch.Queues.FirstOrDefault();
             Assert.NotNull(queue);
             queue.Options.AcknowledgeTimeout = TimeSpan.FromSeconds(3);
@@ -111,10 +111,10 @@ namespace Test.Mq.Statuses
             consumer.ClientId = "consumer";
             await consumer.ConnectAsync("tmq://localhost:" + port);
             Assert.True(consumer.IsConnected);
-            TwinoResult joined = await consumer.Channels.Join("ch-round", true);
+            TwinoResult joined = await consumer.Channels.Join("rr-a", true);
             Assert.Equal(TwinoResultCode.Ok, joined.Code);
 
-            TwinoResult ack = await producer.Queues.Push("ch-round", MessageA.ContentType, "Hello, World!", true);
+            TwinoResult ack = await producer.Queues.Push("rr-a", MessageA.ContentType, "Hello, World!", true);
             Assert.Equal(queueAckIsActive, ack.Code == TwinoResultCode.Ok);
         }
     }
