+++ src/Tests/Test.Queues/Test.Queues.csproj
@@ -18,6 +18,7 @@
       <ProjectReference Include="..\..\Twino.MQ.Data\Twino.MQ.Data.csproj" />
       <ProjectReference Include="..\..\Twino.MQ\Twino.MQ.csproj" />
       <ProjectReference Include="..\..\Twino.Protocols.TMQ\Twino.Protocols.TMQ.csproj" />
+      <ProjectReference Include="..\Test.Common\Test.Common.csproj" />
       <ProjectReference Include="..\Test.Mq\Test.Mq.csproj" />
     </ItemGroup>
