+++ src/Twino.Mq.sln
@@ -15,8 +15,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twino.Protocols.TMQ", "Twin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twino.Client.TMQ", "Twino.Client.TMQ\Twino.Client.TMQ.csproj", "{A75CABEB-197E-457A-AAC9-4C59D49DDF68}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.Mq", "Tests\Test.Mq\Test.Mq.csproj", "{44BCC806-180F-42F7-A848-D3623356D318}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twino.MQ.Data", "Twino.MQ.Data\Twino.MQ.Data.csproj", "{45017EA4-63E1-49BF-8B53-25A45ADF8FCD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.Common", "Tests\Test.Common\Test.Common.csproj", "{66F442DC-EC31-4BD9-B124-64D343B45A6A}"
@@ -61,10 +59,6 @@ Global
 		{A75CABEB-197E-457A-AAC9-4C59D49DDF68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A75CABEB-197E-457A-AAC9-4C59D49DDF68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A75CABEB-197E-457A-AAC9-4C59D49DDF68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{44BCC806-180F-42F7-A848-D3623356D318}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{44BCC806-180F-42F7-A848-D3623356D318}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{44BCC806-180F-42F7-A848-D3623356D318}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{44BCC806-180F-42F7-A848-D3623356D318}.Release|Any CPU.Build.0 = Release|Any CPU
 		{45017EA4-63E1-49BF-8B53-25A45ADF8FCD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{45017EA4-63E1-49BF-8B53-25A45ADF8FCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45017EA4-63E1-49BF-8B53-25A45ADF8FCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -115,7 +109,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{3A3A6CBA-E31C-4E63-A06A-E0FF5DFE1619} = {2BAA30B9-6118-4C33-BD6B-AC6A5827EC7E}
-		{44BCC806-180F-42F7-A848-D3623356D318} = {41761E68-E2ED-4BFE-BBE1-D294A3967EB4}
 		{66F442DC-EC31-4BD9-B124-64D343B45A6A} = {41761E68-E2ED-4BFE-BBE1-D294A3967EB4}
 		{E7ADC6AE-2393-4422-87E7-1CD9A9B4B124} = {41761E68-E2ED-4BFE-BBE1-D294A3967EB4}
 		{F7C4DC6A-01D7-4D1A-893D-58DB39027759} = {41761E68-E2ED-4BFE-BBE1-D294A3967EB4}
